DNS-Exploration-by-python
Use Python to perform reconnaissance and gain information about a target organization using the domain name system or DNS.
DNS is essentially the phonebook of the internet.
we can learn about the IP ranges that belong to a particular organization, the types of systems that are at particular IP addresses etc. And this information is invaluable for planning an attack or a testing exercise against a particular organization.